Add ridesharing for distributed scenario + tests
