+++ source/jormungandr/tests/routing_tests_experimental.py
@@ -26,6 +26,7 @@
 # IRC #navitia on freenode
 # https://groups.google.com/d/forum/navitia
 # www.navitia.io
+
 from __future__ import absolute_import, print_function, unicode_literals, division
 from .tests_mechanism import config, NewDefaultScenarioAbstractTestFixture
 from .journey_common_tests import *
@@ -108,12 +109,14 @@ class TestJourneysDistributed(JourneyCommon, DirectPath, JourneyMinBikeMinCar, N
             "bike": instance.bike_speed,
             "car": instance.car_speed,
             "bss": instance.bss_speed,
+            "ridesharing": instance.car_no_park_speed,
         }
         request = {
             "walking_speed": instance.walking_speed,
             "bike_speed": instance.bike_speed,
             "car_speed": instance.car_speed,
             "bss_speed": instance.bss_speed,
+            "car_no_park_speed": instance.car_no_park_speed,
         }
         resp = instance.get_street_network_routing_matrix([origin], [destination],
                                                           mode, max_duration, request, **kwargs)
@@ -164,12 +167,55 @@ class TestDistributedJourneysWithPtref(JourneysWithPtref, NewDefaultScenarioAbst
 class TestDistributedOnBasicRouting(OnBasicRouting, NewDefaultScenarioAbstractTestFixture):
     @skip("temporarily disabled")
     def test_isochrone(self):
-        super(TestExperimentalOnBasicRouting, self).test_isochrone()
+        super(TestDistributedOnBasicRouting, self).test_isochrone()
+
 
 @config({"scenario": "distributed"})
 class TestDistributedMinNbJourneys(JourneysMinNbJourneys, NewDefaultScenarioAbstractTestFixture):
     pass
 
+
 @config({"scenario": "distributed"})
 class TestDistributedWithNightBusFilter(JourneysWithNightBusFilter, NewDefaultScenarioAbstractTestFixture):
     pass
+
+
+@config({"scenario": "distributed",
+         'instance_config': {
+             "ridesharing": [
+             {
+                 "class": "jormungandr.scenarios.ridesharing.instant_system.InstantSystem",
+                 "args": {
+                     "service_url": "http://distributed_ridesharing.wtf",
+                     "api_key": "key",
+                     "network": "Super Covoit 3000",
+                     "rating_scale_min": 0,
+                     "rating_scale_max": 5
+                 }
+             }
+         ]}})
+class TestJourneysRidesharingDistributed(JourneysRidesharing, JourneyCommon, DirectPath, JourneyMinBikeMinCar,
+                                         NewDefaultScenarioAbstractTestFixture):
+    def test_best_filtering(self):
+        """
+        This feature is not supported
+        """
+        pass
+
+    def test_journeys_wheelchair_profile(self):
+        """
+        This feature is not supported
+        """
+        pass
+
+    def test_not_existent_filtering(self):
+        """
+        This feature is not supported
+        """
+        pass
+
+    def test_other_filtering(self):
+        """
+        This feature is not supported
+        """
+        pass
