+++ source/jormungandr/tests/routing_tests_new_default.py
@@ -27,12 +27,9 @@
 # https://groups.google.com/d/forum/navitia
 # www.navitia.io
 from __future__ import absolute_import, print_function, unicode_literals, division
-from datetime import timedelta
 from .tests_mechanism import NewDefaultScenarioAbstractTestFixture
 from .tests_mechanism import config
-from .check_utils import *
 from .journey_common_tests import *
-from unittest import skip
 
 '''
 This unit runs all the common tests in journey_common_tests.py along with locals tests added in this
@@ -285,7 +282,7 @@ class TestJourneysNewDefault(JourneyCommon,  DirectPath, JourneyMinBikeMinCar, N
                  }
              }
          ]}})
-class TestJourneysRidesharingNewDefault(JourneyCommon,  DirectPath, JourneyMinBikeMinCar,
+class TestJourneysRidesharingNewDefault(JourneysRidesharing, JourneyCommon, DirectPath, JourneyMinBikeMinCar,
                                         NewDefaultScenarioAbstractTestFixture):
     def test_best_filtering(self):
         """
@@ -310,74 +307,7 @@ class TestJourneysRidesharingNewDefault(JourneyCommon,  DirectPath, JourneyMinBi
         This feature is no longer supported
         """
         pass
-
-    def test_first_ridesharing_last_walking_section_mode(self):
-        query = "journeys?from=0.0000898312;0.0000898312&to=0.00188646;0.00071865&datetime=20120614T075500&"\
-                "first_section_mode[]={first}&last_section_mode[]={last}&debug=true"\
-                .format(first='ridesharing', last='walking')
-        response = self.query_region(query)
-        check_best(response)
-        self.is_valid_journey_response(response, query)
-        assert len(response["journeys"]) == 1
-        assert response["journeys"][0]["type"] == "best"
-        rs_journey = response["journeys"][0]
-        assert "ridesharing" in rs_journey["tags"]
-        assert rs_journey["requested_date_time"] == "20120614T075500"
-        assert rs_journey["departure_date_time"] == "20120614T075500"
-        assert rs_journey["arrival_date_time"] == "20120614T075513"
-        assert rs_journey["distances"]["ridesharing"] == 94
-        assert rs_journey["duration"] == 13
-        assert rs_journey["durations"]["ridesharing"] == rs_journey["duration"]
-        assert rs_journey["durations"]["total"] == rs_journey["duration"]
-        assert 'to_delete' in rs_journey["tags"] # no response provided for ridesharing: to_delete
-        rs_section = rs_journey["sections"][0]
-        assert rs_section["departure_date_time"] == rs_journey["departure_date_time"]
-        assert rs_section["arrival_date_time"] == rs_journey["arrival_date_time"]
-        assert rs_section["duration"] == rs_journey["duration"]
-        assert rs_section["mode"] == "ridesharing"
-        assert rs_section["type"] == "street_network"
-        assert rs_section["id"] # check that id is provided
-        assert rs_section["geojson"]["properties"][0]["length"] == rs_journey["distances"]["ridesharing"]
-        assert (rs_section["geojson"]["coordinates"][0][0] - float(rs_section["from"]["address"]["coord"]["lon"])) < 0.000001
-        assert (rs_section["geojson"]["coordinates"][0][1] - float(rs_section["from"]["address"]["coord"]["lat"])) < 0.000001
-        assert (rs_section["geojson"]["coordinates"][1][0] - float(rs_section["to"]["address"]["coord"]["lon"])) < 0.000001
-        assert (rs_section["geojson"]["coordinates"][1][1] - float(rs_section["to"]["address"]["coord"]["lat"])) < 0.000001
-
-    def test_first_ridesharing_section_mode_forbidden(self):
-        def exec_and_check(query):
-            response, status = self.query_region(query, check=False)
-            check_best(response)
-            assert status == 400
-            assert "message" in response
-            assert "ridesharing" in response['message']
-
-        query = "journeys?from=0.0000898312;0.0000898312&datetime=20120614T075500&"\
-                "first_section_mode[]={first}&last_section_mode[]={last}"\
-                .format(first='ridesharing', last='walking')
-        exec_and_check(query)
-
-        query = "isochrones?from=0.0000898312;0.0000898312&datetime=20120614T075500&"\
-                "first_section_mode[]={first}&last_section_mode[]={last}&max_duration=2"\
-                .format(first='ridesharing', last='walking')
-        exec_and_check(query)
-
-        query = "heat_maps?from=0.0000898312;0.0000898312&datetime=20120614T075500&"\
-                "first_section_mode[]={first}&last_section_mode[]={last}&max_duration=2"\
-                .format(first='ridesharing', last='walking')
-        exec_and_check(query)
-
-    def test_free_radius_from(self):
-        """
-        This feature is not supported for this scenario
-        """
-        pass
-
-    def test_free_radius_to(self):
-        """
-        This feature is not supported for this scenario
-        """
-        pass
-
+    
 
 @config({"scenario": "new_default"})
 class TestNewDefaultJourneysWithPtref(JourneysWithPtref, NewDefaultScenarioAbstractTestFixture):
